+++ packages/dependency-metatx/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Add _msgSender and _msgData so no contract that inherit doesn't need to implement it
+- Add a check that the new trusted forwarder is not the same already set.
+
 ## [0.0.2] - 2023-09-01
 
 ### Added